+++ src/Common/TranslatableFields/FieldType.php
@@ -6,10 +6,11 @@ namespace Thinktomorrow\Chief\Common\TranslatableFields;
 
 class FieldType
 {
-    const INPUT = 'input';  // oneliner text (input)
-    const TEXT  = 'text';   // Plain text (textarea)
-    const HTML  = 'html';   // Html text (wysiwyg)
-    const DATE  = 'date';   // Timestamp input
+    const INPUT  = 'input';   // oneliner text (input)
+    const TEXT   = 'text';    // Plain text (textarea)
+    const DATE   = 'date';    // Timestamp input
+    const HTML   = 'html';    // Html text (wysiwyg)
+    const SELECT = 'select';  // Select options
 
     /**
      * @var string
@@ -18,7 +19,7 @@ class FieldType
 
     public function __construct(string $type)
     {
-        if (!in_array($type, [static::INPUT, static::TEXT, static::HTML, static::DATE])) {
+        if (!in_array($type, [static::INPUT, static::TEXT, static::HTML, static::SELECT, static::DATE])) {
             throw new \Exception('Invalid type identifier given ['.$type.'].');
         }