+++ docs/CHANGELOG.md
@@ -11,6 +11,11 @@ All Notable changes to the `chief` application template will be documented in th
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## NEXT - 2018-08-13
+
+### Fixed
+- squanto permission middleware
+
 ## NEXT - 2018-08-09
 
 ### Added
@@ -19,7 +24,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## NEXT - 2018-07-27
 
 ### Added
-- updated docs
+- updated docsp te  
 
 ### Fixed
 - Dont show hidden_in_menu pages in the internal page dropdown for menuitems: this had weird behaviour as linking a hidden menu would hide the menuitem everywhere. Even in the admin.